auto-improve: Merge triplicated `_mk_result` SDK ResultMessage builder

Refs damien-robotsix/robotsix-cai#1322
